moving S3Specification to utils
