+++ sentinelhub/api/utils.py
@@ -2,10 +2,11 @@
 Module implementing some common utility functions
 """
 from enum import Enum
-from typing import Any, Dict, Type
+from typing import Any, Dict, Optional, Type, TypedDict
 
 from dataclasses_json import LetterCase
 from dataclasses_json import config as dataclass_config
+from typing_extensions import NotRequired
 
 from ..geometry import Geometry
 from ..time_utils import parse_time, serialize_time
@@ -48,3 +49,36 @@ def _update_other_args(dict1: Dict[str, Any], dict2: Dict[str, Any]) -> None:
 def remove_undefined(payload: dict) -> dict:
     """Takes a dictionary and removes keys without value"""
     return {name: value for name, value in payload.items() if value is not None}
+
+
+class S3Specification(TypedDict):
+    """Specification of a S3 path."""
+
+    url: str
+    accessKey: str
+    secretAccessKey: str
+    region: NotRequired[str]
+
+
+class AccessSpecification(TypedDict):
+    """Specification of a S3 input or output."""
+
+    s3: S3Specification
+
+
+def s3_specification(
+    url: str, access_key: str, secret_access_key: str, region: Optional[str] = None
+) -> AccessSpecification:
+    """A helper method to build a dictionary used for specifying S3 paths
+
+    :param url: A URL pointing to an S3 bucket or an object in an S3 bucket.
+    :param access_key: AWS access key that allows programmatic access to the S3 bucket specified in the `url` field.
+    :param secret_access_key: AWS secret access key which must correspond to the AWS access key.
+    :param region: The region where the S3 bucket is located. If omitted, the region of the Sentinel Hub deployment
+        that the request is submitted to is assumed.
+    :return: A dictionary of S3 specifications used by the Batch Statistical API
+    """
+    s3_access: S3Specification = {"url": url, "accessKey": access_key, "secretAccessKey": secret_access_key}
+    if region is not None:
+        s3_access["region"] = region
+    return {"s3": s3_access}
